refactored the pie chart into another js file. should be into an object (OOP), not a function
